[BE] 체크리스트 작성 API를 추가 구현한다. (#310)

+++ backend/bang-ggood/src/main/java/com/bang_ggood/room/dto/response/SelectedRoomResponse.java
@@ -4,7 +4,7 @@ import com.bang_ggood.checklist.domain.Checklist;
 
 public record SelectedRoomResponse(String roomName, Integer deposit, Integer rent, Integer contractTerm, Integer floor,
                                    String address, String station, Integer walkingTime, String realEstate,
-                                   String type, Integer size, String floorLevel, String structure) {
+                                   String type, Double size, String floorLevel, String structure) {
 
     public static SelectedRoomResponse of(Checklist checklist) {
         return new SelectedRoomResponse(checklist.getRoomName(), checklist.getDeposit(), checklist.getRent(),
